refactor : 결제 트랜잭션 분리

+++ Mavis-Domain/src/main/java/com/mavis/domain/domains/order/domain/Payment.java
@@ -28,7 +28,7 @@ public class Payment extends BaseEntity {
     @Builder.Default
     private PaymentMethod method = PaymentMethod.DEFAULT;
 
-    private Long totalAmount;
+    private int totalAmount;
 
     private Long balanceAmount;
