+++ Mavis-Infrastructure/src/main/java/com/mavis/infrastructure/outer/api/tosspayments/dto/PaymentsResponse.java
@@ -12,7 +12,7 @@ public record PaymentsResponse(
         String mId,
         String currency,
         TossPaymentMethod method,
-        Long totalAmount,
+        int totalAmount,
         Long balanceAmount,
         PaymentsStatus status,
         ZonedDateTime requestedAt,
